chore: fix node import
