+++ packages/blocks/src/bookmark-block/index.ts
@@ -1,3 +1,3 @@
 export * from './bookmark-block.js';
 export * from './bookmark-model.js';
-export { getBookmarkInitialProps } from './utils.js';
+export { getBookmarkInitialProps } from './components/bookmark-create-modal.js';
